refactor(user/entity): ♻️ Replace the constructors by lombok annotations
